requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
--trusted-host pypi.python.org
--trusted-host pypi.org
--trusted-host files.pythonhosted.org

amqp==5.1.1
    # via kombu
asgiref==3.5.2
    # via django
async-timeout==4.0.2
    # via redis
attrs==22.1.0
    # via pytest
autopep8==1.7.0
    # via -r requirements.in
billiard==3.6.4.0
    # via celery
celery==5.2.7
    # via
    #   django-celery-beat
    #   django-celery-results
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
colorama==0.4.6
    # via
    #   click
    #   pytest
coverage==6.5.0
    # via -r requirements.in
deprecated==1.2.13
    # via redis
django==4.1.3
    # via
    #   -r requirements.in
    #   django-celery-beat
    #   django-linear-migrations
django-analytical==3.1.0
    # via -r requirements.in
django-celery-beat==2.4.0
    # via -r requirements.in
django-celery-results==2.4.0
    # via -r requirements.in
django-crispy-forms==1.14.0
    # via -r requirements.in
django-linear-migrations==2.5.1
    # via -r requirements.in
django-timezone-field==5.0
    # via django-celery-beat
exceptiongroup==1.0.3
    # via pytest
flake8==5.0.4
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
hiredis==2.0.0
    # via -r requirements.in
iniconfig==1.1.1
    # via pytest
kombu==5.2.4
    # via celery
mccabe==0.7.0
    # via flake8
packaging==21.3
    # via
    #   pytest
    #   redis
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.31
    # via click-repl
psycopg2-binary==2.9.5
    # via -r requirements.in
pycodestyle==2.9.1
    # via
    #   autopep8
    #   flake8
pyflakes==2.5.0
    # via flake8
pyparsing==3.0.9
    # via packaging
pytest==7.2.0
    # via -r requirements.in
python-crontab==2.6.0
    # via django-celery-beat
python-dateutil==2.8.2
    # via python-crontab
python-dotenv==0.21.0
    # via -r requirements.in
pytz==2022.5
    # via
    #   celery
    #   django-timezone-field
redis==4.3.4
    # via -r requirements.in
six==1.16.0
    # via
    #   click-repl
    #   python-dateutil
sqlparse==0.4.3
    # via django
toml==0.10.2
    # via autopep8
tomli==2.0.1
    # via pytest
tzdata==2022.5
    # via
    #   django
    #   django-celery-beat
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
wrapt==1.14.1
    # via deprecated

# The following packages are considered to be unsafe in a requirements file:
# setuptools
